Refactor 'ligacao' parameter type from int to string across multiple files; update validation rules and tests accordingly. Add QR code image generation to DocumentoService and related tests.

+++ app/Http/Controllers/Api/V1/LinhaDigitavelController.php
@@ -20,7 +20,7 @@ class LinhaDigitavelController
         $dados = $request->validated();
 
         $linhaDigitavel = $this->linhaDigitavelService->gerar(
-            ligacao:    (int)   $dados['ligacao'],
+            ligacao:    $dados['ligacao'],
             referencia: (string) $dados['referencia'],
             vencimento: (string) $dados['vencimento'],
             valor:      (float)  $dados['valor'],
